+++ source/consumers/Spectrum.h
@@ -11,7 +11,7 @@
 #include <core/Consumer.h>
 #include <consumers/add_ons/periodic_trigger.h>
 #include <consumers/add_ons/recent_rate.h>
-#include <consumers/add_ons/filter_block.h>
+#include <consumers/add_ons/FilterBlock.h>
 
 namespace DAQuiri {
