+++ source/consumers/add_ons/FilterBlock.h
