+++ tests/consumers/add_ons/FilterBlockTest.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include <consumers/add_ons/filter_block.h>
+#include <consumers/add_ons/FilterBlock.h>
 
 using namespace DAQuiri;
